+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Tons of commands:
   - The internal system to handle commands has been redesigned.
   - Using the `/help` command will trigger a new in-application help modal.
+  - Add the `/user` command to open a user viewer card without any click or having to find a user message first.
   - Add support for the `/block` & `/unblock` commands.
   - Add the `/purge` command which is an alias of the `/timeout user 1`.